requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
exceptiongroup==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
-e git+https://github.com/aFastHero/langchain.git@eccc62c17a7d9c1813e03a7f557bd50496741d31#egg=langchain
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
SQLAlchemy==2.0.10
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.16.1
yarl==1.9.2
